integer overflow

+++ t4-reduce.py
@@ -3,14 +3,14 @@
 import sys
 
 prev_bigram = ""
-value_total = 0
+value_total = 0.0
 bigram=""
 
 for line in sys.stdin:          # For ever line in the input from stdin
     try:
         line = line.strip()         # Remove trailing characters
         bigram, value = line.split("\t", 1)
-        value = int(value)
+        value = float(value)
         # Remember that Hadoop sorts map output by key reducer takes these keys sorted
         if prev_bigram == bigram:
             value_total += value
